more debugging
git-svn-id: svn://localhost/ardour2/trunk@1045 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/automation_line.cc
@@ -744,7 +744,7 @@ AutomationLine::determine_visible_control_points (ALPoints& points)
 		*/
 
 		this_rx = (uint32_t) rint (tx);
-      		this_ry = (unsigned long) rint (ty); 
+      		this_ry = (uint32_t) rint (ty); 
  
  		if (view_index && pi != npoints && (this_rx == prev_rx) && (this_ry == prev_ry) || 
 		    ((this_rx - prev_rx) < (box_size + 2))) {
